requirements/base.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
attrs==25.4.0
cffi==2.0.0
celery==5.6.2
redis==7.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
django-health-check==3.20.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-debug-toolbar
django-environ==0.11.2
django_extensions
drf-spectacular==0.29.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mysqlclient==2.2.7
packaging==25.0
pycparser==2.23
PyJWT==2.10.1
Pillow==12.1.0
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
